Fix socket.destroySoon() problem with http module.
Problems:

If http request is not keepAlive, upon responseEnd, destroySoon
is used to close the response stream. destroySoon does not exist
for SocksClientSocket.

For the above condition, the http.js code expects the socket
to be not writable after returning from destroySoon.

Fixes:

Add destroySoon method to SocksClientSocket
destroySoon sets writable to false.